Added Database Handler (semi-functional DB)
Updates to Account and Transaction (Using username and accountName as Keys rather than ID).

Updates to RegisterActivity - should query the DB and add a new user if one with the same username does not already exist.

Updates to LoginActivity - should query the DB and login if the user is found

Updates to UserHomeActivity - should display user's name, user's accounts, and allow them to click the Account in the list or create a new account.

Added CreateAccount activity - a way to add accounts to the database/ user

Updates to AccountActivity - should change the title and available to the Account's name and balance respectively.

Notes: If you create a new Account make sure the name is 100% unique.